Move EmptyCartView to its own file
